backend/pyproject.toml
[project]
name = "buywander-backend"
version = "0.1.0"
description = "BuyWander Deal Hunter Backend API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlmodel>=0.0.22",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    "pyjwt>=2.9.0",
    "bcrypt>=4.2.0",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.9.2",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.2",
    "apscheduler>=3.10.4",
    "redis>=5.2.0",
    "pandas>=2.1.4",
    "requests>=2.31.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.7.4",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]  # line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint.isort]
known-first-party = ["models", "database", "auth_utils"]
